uninitialized variables checked. Also change two if/else if functions for switch
git-svn-id: http://ltt.polymtl.ca/svn@684 04897980-b3bd-0310-b5e0-8ef037075253

+++ ltt/branches/poly/ltt/trace.h
@@ -151,7 +151,7 @@ void getDataEndianType(LttArchSize * size, LttArchEndian * endian);
 
 /* get an integer number */
 
-int getIntNumber(int size1, void *evD);
+gint64 getIntNumber(int size1, void *evD);
 
 
 /* get the node name of the system */
